+++ app/src/main/java/com/example/dancognitionapp/assessment/bart/data/Balloon.kt
@@ -1,4 +1,4 @@
-package com.example.dancognitionapp.bart.data
+package com.example.dancognitionapp.assessment.bart.data
 
 data class Balloon(
     val listPosition: Int,
