+++ app/src/main/java/com/example/dancognitionapp/utils/LandscapePreview.kt
@@ -1,4 +1,4 @@
-package com.example.dancognitionapp.ui
+package com.example.dancognitionapp.utils
 
 import androidx.compose.ui.tooling.preview.Preview
