+++ app/src/main/java/com/example/dancognitionapp/utils/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.dancognitionapp.ui.theme
+package com.example.dancognitionapp.utils.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
