+++ app/src/main/java/com/example/dancognitionapp/utils/widget/OptionCard.kt
@@ -1,4 +1,4 @@
-package com.example.dancognitionapp.ui.widget
+package com.example.dancognitionapp.utils.widget
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.IdRes
@@ -23,7 +23,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.example.dancognitionapp.ui.landing.LandingDestination
+import com.example.dancognitionapp.landing.LandingDestination
 
 @Composable
 fun OptionCard(
